Move ListUtils to the root of the project
