+++ app/src/main/java/com/twofauth/android/ListUtils.java
@@ -1,9 +1,9 @@
-package com.twofauth.android.main_activity;
+package com.twofauth.android;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Collection;
+import java.util.Arrays;
 import java.util.List;
 
 public class ListUtils {
@@ -11,6 +11,13 @@ public class ListUtils {
         dest.clear();
         if (source != null) {
             dest.addAll(source);
+        }
+    }
+
+    public static <E> void setItems(@NotNull final List<E> dest, @Nullable final E[] source) {
+        dest.clear();
+        if (source != null) {
+            dest.addAll(Arrays.asList(source));
         }
     }
 }
